[sql] Add ship and user table
Add the ship and user table to the create table query.

There is a one to many relationship with the table user and ship. A user
can own multiple ships, but a ship can only belong to one user.

Because of this, a foreign key constrain is added to the ship.